+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Display race traits on class features tab
+- Filters for class features (traits & auto features)
 
 ## [2.7.0] - 2019-05-01
 ### Added